Abstract spring doc extension to nacos-swagger module.
