+++ swagger/src/main/java/com/alibaba/nacos/springdoc/cache/LocaleThreadLocalHolder.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 1999-2025 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.nacos.springdoc.cache;
+
+import java.util.Locale;
+
+/**
+ * spring doc locale thread local holder.
+ *
+ * @author xiweng.yy
+ */
+public class LocaleThreadLocalHolder {
+    
+    private static final ThreadLocal<Locale> LANGUAGE_LOCALE = ThreadLocal.withInitial(Locale::getDefault);
+    
+    public static Locale getLocale() {
+        return LANGUAGE_LOCALE.get();
+    }
+    
+    public static void setLocale(Locale locale) {
+        LANGUAGE_LOCALE.set(locale);
+    }
+}
